requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
pillow
opencv-python
openpyxl
scipy
joblib
